+++ Learns.Model/Interfaces/IInstituteRepository.cs
@@ -9,7 +9,7 @@ namespace Learns.Model.Interfaces
     public interface IInstituteRepository
     {
         Task<IEnumerable<Institute>> GetInstitutes();
-        Task<Institute> GetInstituteByID(int ID);
+        Task<Institute> GetInstituteByID(long ID);
         Task<Institute> InsertInstitute(Institute objInstitute);
         Task<Institute> UpdateInstitute(Institute objInstitute);
         bool DeleteInstitute(long ID);
